refactor(chess-game): 클래스 네이밍 변경. 체스 컨트롤러 -> 체스 게임
