+++ src/main/java/chess/domain/ChessGame.java
@@ -1,7 +1,5 @@
-package chess.controller;
+package chess.domain;
 
-import chess.domain.ChessExecution;
-import chess.domain.MoveResult;
 import chess.domain.board.Board;
 import chess.domain.board.BoardFactory;
 import chess.dto.BoardDto;
@@ -9,16 +7,27 @@ import chess.dto.ScoreDto;
 import chess.view.InputView;
 import chess.view.OutputView;
 
-public class ChessController {
+public class ChessGame {
+    private static final InputView inputView = InputView.getInstance();
+    private static final OutputView outputView = OutputView.getInstance();
+    private static final String COMMAND_DELIMITER = " ";
+
+    private final Board board;
+
+    public ChessGame() {
+        this.board = BoardFactory.newInstance();
+    }
+
     public void start() {
-        OutputView outputView = OutputView.getInstance();
-        InputView inputView = InputView.getInstance();
-        Board board = BoardFactory.newInstance();
         outputView.printBoard(BoardDto.from(board));
 
-        String[] command = inputView.scanCommand().split(" ");
+        String[] command = inputView.scanCommand().split(COMMAND_DELIMITER);
         ChessExecution execution = ChessExecution.from(command[0]);
 
+        play(board, command, execution);
+    }
+
+    private void play(Board board, String[] command, ChessExecution execution) {
         while (execution != ChessExecution.END) {
             if (execution == ChessExecution.MOVE) {
                 MoveResult move = board.move(command[1], command[2]);
@@ -34,7 +43,7 @@ public class ChessController {
                 outputView.printScore(ScoreDto.from(board.getScore()));
             }
 
-            command = inputView.scanCommand().split(" ");
+            command = inputView.scanCommand().split(COMMAND_DELIMITER);
             execution = ChessExecution.from(command[0]);
         }
     }
